requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.0.0
python-dotenv==1.0.0
requests==2.31.0
